refactor(precompiles): flatten token precompile modules
