+++ crates/common/precompiles/src/b20/mod.rs
@@ -1,6 +1,8 @@
 //! `B20Token` native precompile — the core B-20 token implementation.
 
+mod abi;
 mod dispatch;
+pub use abi::IB20;
 
 mod precompile;
 pub use precompile::B20TokenPrecompile;
