+++ crates/common/precompiles/src/b20/precompile.rs
@@ -4,7 +4,7 @@ use alloy_evm::precompiles::DynPrecompile;
 use alloy_primitives::Address;
 
 use super::{B20Token, storage::B20TokenStorage};
-use crate::{macros::base_precompile, token::PolicyHandle};
+use crate::{PolicyHandle, macros::base_precompile};
 
 /// Entry point for the `B20Token` precompile.
 ///
