+++ crates/common/precompiles/src/common/ops/mintable.rs
@@ -2,10 +2,7 @@ use alloy_primitives::{Address, B256, U256};
 use alloy_sol_types::SolEvent;
 use base_precompile_storage::{BasePrecompileError, Result};
 
-use crate::token::{
-    IB20,
-    common::{Token, TokenAccounting},
-};
+use crate::{IB20, Token, TokenAccounting};
 
 /// Token minting operations.
 ///
