+++ crates/common/precompiles/src/common/token_accounting.rs
