+++ crates/common/precompiles/src/factory/mod.rs
@@ -1,6 +1,8 @@
 //! `TokenFactory` native precompile — creates B-20 tokens at deterministic prefix-encoded addresses.
 
+mod abi;
 mod dispatch;
+pub use abi::ITokenFactory;
 
 mod precompile;
 pub use precompile::TokenFactoryPrecompile;
