+++ crates/common/precompiles/src/factory/variant.rs
