+++ crates/common/precompiles/src/lib.rs
@@ -23,11 +23,19 @@ mod bn254_pair;
 
 mod bls12_381;
 
-mod token;
-pub use token::{
-    B20Token, B20TokenPrecompile, B20TokenStorage, Burnable, CAPABILITY_CAP_MUTABLE,
-    CAPABILITY_PAUSABLE, Configurable, IB20, IPolicyRegistry, ITokenFactory, Mintable,
-    POLICY_REGISTRY_ADDRESS, Pausable, Permittable, Policy, PolicyHandle, PolicyRegistryEvm,
-    Redeemable, Token, TokenAccounting, TokenFactory, TokenFactoryPrecompile, TokenVariant,
-    Transferable,
+mod common;
+pub use common::{
+    Burnable, CAPABILITY_CAP_MUTABLE, CAPABILITY_PAUSABLE, Configurable, Mintable, Pausable,
+    Permittable, Policy, Redeemable, Token, TokenAccounting, Transferable,
+};
+
+mod b20;
+pub use b20::{B20Token, B20TokenPrecompile, B20TokenStorage, IB20};
+
+mod factory;
+pub use factory::{ITokenFactory, TokenFactory, TokenFactoryPrecompile, TokenVariant};
+
+mod policy_registry;
+pub use policy_registry::{
+    IPolicyRegistry, POLICY_REGISTRY_ADDRESS, PolicyHandle, PolicyRegistryEvm,
 };
