+++ crates/common/precompiles/src/policy_registry/mod.rs
@@ -1,5 +1,8 @@
 //! `PolicyRegistry` native precompile — global singleton transfer-policy registry for B-20 tokens.
 
+mod abi;
+pub use abi::IPolicyRegistry;
+
 mod dispatch;
 
 mod evm;
