--- crates/common/precompiles/src/token/abi/mod.rs
@@ -1,10 +1,0 @@
-//! ABI types for the token precompile domain.
-
-mod b20;
-pub use b20::IB20;
-
-mod factory;
-pub use factory::ITokenFactory;
-
-mod policy_registry;
-pub use policy_registry::IPolicyRegistry;
